Criação do controlador CartController e views

+++ VShop.Web/Models/CartHeaderViewModel.cs
@@ -6,5 +6,5 @@ public class CartHeaderViewModel
     public string UserId { get; set; } = string.Empty;
     public string CouponCode { get; set; } = string.Empty;
 
-    public double TotalAmount { get; set; } = 0.00d;
+    public decimal TotalAmount { get; set; } = 0.00m;
 }
